파일, 멤버, 게임 패키지 구조 변경 (#347)
* chore: 패키지 구조 재설정

* chore: h2 -> mysql로 DBMS 변경함에 따라 기본 키 생성전략을 IDENTITY로 변경

* refactor: @Column 어노테이션 삭제처리

* feat: 기존에 중복되는 FileType 클래스명을 FileFormat으로 변경

* feat: 기존에 중복되는 FileType 클래스명을 FileFormat으로 변경

* refactor: 기존 파일 필드 네이밍 변경 뒤 충돌나는 것들 수정

* 톡픽 관련 엔티티 클래스 작성 (#346)

* feat: 톡픽 관련 엔티티 클래스 작성

* feat: 톡픽 좋아요 관련 엔티티 추가

* chore: Bookmark 패키지 구조 개선

* chore: conflict merge

* chore: h2 -> mysql로 DBMS 변경함에 따라 기본 키 생성전략을 IDENTITY로 변경

* refactor: @Column 어노테이션 삭제처리

* feat: 기존에 중복되는 FileType 클래스명을 FileFormat으로 변경

* feat: 기존에 중복되는 FileType 클래스명을 FileFormat으로 변경

* refactor: 기존 파일 필드 네이밍 변경 뒤 충돌나는 것들 수정

* chore: conflict merge

---------
