+++ src/main/java/balancetalk/authmail/dto/EmailRequest.java
@@ -1,0 +1,20 @@
+package balancetalk.authmail.dto;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+public class EmailRequest {
+
+    @NotNull
+    @Size(max = 30)
+    @Email(regexp = "^[a-zA-Z0-9._%+-]{1,20}@[a-zA-Z0-9.-]{1,10}\\.[a-zA-Z]{2,}$")
+    @Schema(description = "인증 번호를 받을 이메일 주소", example = "test1234@naver.com")
+    private String email;
+}
