+++ src/main/java/balancetalk/bookmark/domain/BookmarkType.java
@@ -1,0 +1,5 @@
+package balancetalk.bookmark.domain;
+
+public enum BookmarkType {
+    GAME
+}
