+++ src/main/java/balancetalk/file/domain/FileRepository.java
@@ -1,0 +1,8 @@
+package balancetalk.file.domain;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FileRepository extends JpaRepository<File, Long> {
+    Optional<File> findByStoredName(String storedName);
+}
