+++ src/main/java/balancetalk/file/domain/FileType.java
@@ -1,0 +1,5 @@
+package balancetalk.file.domain;
+
+public enum FileType {
+    TALK_PICK, GAME
+}
