+++ src/main/java/balancetalk/like/domain/Like.java
@@ -1,7 +1,7 @@
 package balancetalk.like.domain;
 
 import balancetalk.global.common.BaseTimeEntity;
-import balancetalk.module.member.domain.Member;
+import balancetalk.member.domain.Member;
 import balancetalk.talkpick.domain.TalkPick;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
