+++ src/main/java/balancetalk/member/domain/Role.java
@@ -1,0 +1,5 @@
+package balancetalk.member.domain;
+
+public enum Role {
+    ADMIN, USER
+}
