--- src/main/java/balancetalk/module/authmail/dto/EmailVerification.java
@@ -1,23 +1,0 @@
-package balancetalk.module.authmail.dto;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-public class EmailVerification {
-
-    @NotNull
-    @Size(max = 30)
-    @Email(regexp = "^[a-zA-Z0-9._%+-]{1,20}@[a-zA-Z0-9.-]{1,10}\\.[a-zA-Z]{2,}$")
-    @Schema(description = "인증 번호를 검증할 이메일 주소", example = "test1234@naver.com")
-    private String email;
-
-    @Schema(description = "인증 번호", example = "4f7dfb")
-    private String verificationCode;
-}
