--- src/main/java/balancetalk/module/member/domain/Role.java
@@ -1,5 +1,0 @@
-package balancetalk.module.member.domain;
-
-public enum Role {
-    ADMIN, USER
-}
