+++ src/test/java/balancetalk/module/member/application/MemberServiceTest.java
@@ -4,11 +4,14 @@ import balancetalk.global.exception.BalanceTalkException;
 import balancetalk.global.exception.ErrorCode;
 import balancetalk.global.jwt.JwtTokenProvider;
 import balancetalk.global.redis.application.RedisService;
-import balancetalk.module.file.domain.File;
-import balancetalk.module.file.domain.FileRepository;
-import balancetalk.module.member.domain.Member;
-import balancetalk.module.member.domain.MemberRepository;
-import balancetalk.module.member.dto.*;
+import balancetalk.member.application.MemberService;
+import balancetalk.member.dto.JoinRequest;
+import balancetalk.member.dto.LoginRequest;
+import balancetalk.member.dto.MemberResponse;
+import balancetalk.file.domain.File;
+import balancetalk.file.domain.FileRepository;
+import balancetalk.member.domain.Member;
+import balancetalk.member.domain.MemberRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.AfterEach;
